# api-swagger.yaml
openapi: 3.0.0
info:
  description: Edu attend API
  version: '1.0.0'
  title: Edu attend API
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: trhoangthien1999@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:4000'
...
paths:
...
  /grades:
    post:
      tags:
        - grades
      summary: Create grade
      operationId: createGrade
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GradeBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Create grade success'
                  result:
                    $ref: '#/components/schemas/Grade'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '422':
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
  /parents:
    get:
      tags:
        - parents
      summary: Get all parents
      operationId: getParents
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get all parents success'
                  data:
                    type: object
                    properties:
                      parents:
                        type: array
                        items:
                          $ref: '#/components/schemas/Parent'
        '401':
          description: Unauthorized
    post:
      tags:
        - parents
      summary: Create parent
      operationId: createParent
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parent'
        '401':
          description: Unauthorized
  /parents/{id}:
    get:
      tags:
        - parents
      summary: Get parent detail
      operationId: getParentDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parent'
        '401':
          description: Unauthorized
...
  /parents/{id}/students:
    get:
      tags:
        - parents
      summary: Get linked students
      operationId: getLinkedStudents
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get linked students success'
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/Student'
        '401':
          description: Unauthorized

    patch:
      tags:
        - parents
      summary: Update parent
      operationId: updateParent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParentBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parent'
        '401':
          description: Unauthorized
    delete:
      tags:
        - parents
      summary: Delete parent
      operationId: deleteParent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Delete parent success'
        '401':
          description: Unauthorized
...
  /classes:
    get:
      tags:
        - classes
      summary: Get all classes
      operationId: getClasses
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: query
          description: Filter by class name
          required: false
          schema:
            type: string
        - name: code
          in: query
          description: Filter by class code
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get classes success'
                  result:
                    type: object
                    properties:
                      classes:
                        type: array
                        items:
                          $ref: '#/components/schemas/Class'
                      total:
                        type: integer
                        example: 100
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      total_pages:
                        type: integer
                        example: 10
    post:
      tags:
        - classes
      summary: Create class
      operationId: createClass
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Create class success'
                  result:
                    $ref: '#/components/schemas/Class'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '422':
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /classes/{id}:
    get:
      tags:
        - classes
      summary: Get class detail
      operationId: getClassDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: MongoId
            example: '69218ec683ccdc34f38abfbd'
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get class detail success'
                  result:
                    $ref: '#/components/schemas/Class'
        '401':
          description: Unauthorized
    patch:
      tags:
        - classes
      summary: Update class
      operationId: updateClass
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: MongoId
            example: '69218ec683ccdc34f38abfbd'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Class'
        '401':
          description: Unauthorized
        '422':
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
    delete:
      tags:
        - classes
      summary: Delete class
      operationId: deleteClass
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: MongoId
            example: '69218ec683ccdc34f38abfbd'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Delete class success'
        '401':
          description: Unauthorized
...
  /roles:
    get:
      tags:
        - roles
      summary: Get all roles
      operationId: getRoles
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get all roles success'
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'
        '401':
          description: Unauthorized
  /roles/{id}:
    get:
      tags:
        - roles
      summary: Get role detail
      operationId: getRoleDetail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Unauthorized
...
  /roles/user/{user_id}:
    get:
      tags:
        - roles
      summary: Get user roles
      operationId: getUserRoles
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get user roles successfully'
                  result:
                    type: array
                    items:
                      type: object # UserRole info
        '401':
          description: Unauthorized
...
components:
  schemas:
...
    Unauthorized:
      type: object
      properties:
        message:
          type: string
          example: 'Unauthorized'
...
    ValidationError:
      type: object
      properties:
        message:
          type: string
          example: 'Validation error'
        errors:
          type: object
          properties:
            email:
              type: object
              properties:
                type:
                  type: string
                  example: 'field'
                value:
                  type: string
                  example: 'tranhoangthiendev@gmail.com'
                msg:
                  type: string
                  example: 'Email or password is incorrect'
                path:
                  type: string
                  example: 'email'
                location:
                  type: string
                  example: 'body'
...
    GradeBody:
      type: object
      properties:
        name:
          type: string
          example: '10'
        level:
          type: number
          example: 10
    Grade:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: '69218ec683ccdc34f38abfbd'
        name:
          type: string
          example: 'HCM'
        level:
          type: number
          example: 10
...
    ClassBody:
      type: object
      properties:
        name:
          type: string
          example: 'Class'
        code:
          type: string
          example: 'CLS001'
        grade_id:
          type: string
          format: MongoId
          example: '69218ec683ccdc34f38abfbd'
        school_id:
          type: string
          format: MongoId
          example: '69218ec683ccdc34f38abfbd'
        academic_year_id:
          type: string
          format: MongoId
          example: '69218ec683ccdc34f38abfbd'
    Class:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: '69218ec683ccdc34f38abfbd'
        name:
          type: string
          example: 'Class'
        code:
          type: string
          example: 'CLS001'
        grade_id:
          type: string
          format: MongoId
          example: '69218ec683ccdc34f38abfbd'
        school_id:
          type: string
          format: MongoId
          example: '69218ec683ccdc34f38abfbd'
        academic_year_id:
          type: string
          format: MongoId
          example: '69218ec683ccdc34f38abfbd'
        created_at:
          type: string
          format: ISO8601
          example: '2025-11-22T10:21:58.648Z'
        updated_at:
          type: string
          format: ISO8601
          example: '2025-11-22T10:45:46.131Z'
...
    Student:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
        student_code:
          type: string
        full_name:
          type: string
        class_id:
          type: string
          format: MongoId
        school_id:
          type: string
          format: MongoId
        status:
          type: integer
        created_at:
          type: string
          format: ISO8601
        updated_at:
          type: string
          format: ISO8601
    ParentBody:
      type: object
      required:
        - full_name
        - phone
      properties:
        full_name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        job:
          type: string
    Parent:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
        user_id:
          type: string
          format: MongoId
        full_name:
          type: string
        phone:
          type: string
        students:
          type: array
          items:
            type: string
            format: MongoId
        created_at:
          type: string
          format: ISO8601
        updated_at:
          type: string
          format: ISO8601
...
    Permission:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
        name:
          type: string
        code:
          type: string
        description:
          type: string
        module:
          type: string
        created_at:
          type: string
          format: ISO8601
        updated_at:
          type: string
          format: ISO8601
...
    Role:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
        name:
          type: string
        code:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        created_at:
          type: string
          format: ISO8601
        updated_at:
          type: string
          format: ISO8601
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT